Separate camera rotate from move; Move camera implementations into Core;
