+++ Augmented/Graphics/Camera/GameView3D.cs
@@ -52,20 +52,21 @@ namespace Augmented.Graphics.Camera
         public Task<Unit> Handle(Move3DViewRequest request, CancellationToken cancellationToken)
         {
             Camera.GameUpdateContinuousMovement = request.CameraMovementFlags;
+
             return Unit.Task;
         }
 
         public Task<Unit> Handle(Rotate3DViewRequest request, CancellationToken cancellationToken)
         {
             if (request.XRotation > float.Epsilon)
-                Camera.Move(CameraMovement.RotateDown, request.XRotation);
+                Camera.Rotate(CameraMovement.RotateDown, request.XRotation);
             else if (request.XRotation < float.Epsilon)
-                Camera.Move(CameraMovement.RotateUp, -request.XRotation);
+                Camera.Rotate(CameraMovement.RotateUp, -request.XRotation);
 
             if (request.ZRotation > float.Epsilon)
-                Camera.Move(CameraMovement.RotateLeft, request.ZRotation);
+                Camera.Rotate(CameraMovement.RotateLeft, request.ZRotation);
             else if (request.ZRotation < float.Epsilon)
-                Camera.Move(CameraMovement.RotateRight, -request.ZRotation);
+                Camera.Rotate(CameraMovement.RotateRight, -request.ZRotation);
 
             return Unit.Task;
         }
