+++ Augmented/UserInterface/Input/CameraMovement/ICameraMovement.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.Xna.Framework.Input;
+
+namespace Augmented.UserInterface.Input.CameraMovementSpace
+{
+    public interface ICameraMovement
+    {
+        void MoveCamera(Keys[] keysDown);
+    }
+}
