+++ Augmented/UserInterface/Input/GameViewKeyboardHandler.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Linq;
+
 using Augmented.Messages;
+using Augmented.UserInterface.Input.CameraMovementSpace;
 
 using DavidFidge.MonoGame.Core.Graphics.Camera;
 using DavidFidge.MonoGame.Core.UserInterface;
@@ -13,6 +15,12 @@ namespace Augmented.UserInterface.Input
 {
     public class GameViewKeyboardHandler : BaseKeyboardHandler
     {
+        private readonly ICameraMovement _cameraMovement;
+
+        public GameViewKeyboardHandler(ICameraMovement cameraMovement)
+        {
+            _cameraMovement = cameraMovement;
+        }
         public override void HandleKeyboardKeyDown(Keys[] keysDown, Keys keyInFocus, KeyboardModifier keyboardModifier)
         {
             if (keyInFocus == Keys.Escape)
@@ -21,110 +29,17 @@ namespace Augmented.UserInterface.Input
             if (keyInFocus == Keys.F12)
                 Environment.Exit(0);
 
-            SetCameraContinuousMovement(keysDown);
+            _cameraMovement.MoveCamera(keysDown);
         }
 
         public override void HandleKeyboardKeyLost(Keys[] keysDown, KeyboardModifier keyboardModifier)
         {
-            SetCameraContinuousMovement(keysDown);
+            _cameraMovement.MoveCamera(keysDown);
         }
 
         public override void HandleKeyboardKeysReleased()
         {
             Mediator.Send(new Move3DViewRequest(CameraMovement.None));
-        }
-
-        private void SetCameraContinuousMovement(Keys[] keysDown)
-        {
-            FreeCameraMovement(keysDown);
-        }
-
-        private void StrategyMovement(Keys[] keysDown)
-        {
-            var cameraMovementFlags = CameraMovement.None;
-
-            if (keysDown.Contains(Keys.A) || keysDown.Contains(Keys.Left))
-                cameraMovementFlags |= CameraMovement.PanLeft;
-
-            if (keysDown.Contains(Keys.D) || keysDown.Contains(Keys.Right))
-                cameraMovementFlags |= CameraMovement.PanRight;
-
-            if (keysDown.Contains(Keys.W) || keysDown.Contains(Keys.Up))
-                cameraMovementFlags |= CameraMovement.PanUp;
-
-            if (keysDown.Contains(Keys.S) || keysDown.Contains(Keys.Down))
-                cameraMovementFlags |= CameraMovement.PanDown;
-
-            if (keysDown.Contains(Keys.Q))
-                cameraMovementFlags |= CameraMovement.RotateLeft;
-
-            if (keysDown.Contains(Keys.E))
-                cameraMovementFlags |= CameraMovement.RotateRight;
-
-            if (keysDown.Contains(Keys.PageUp) || keysDown.Contains(Keys.R))
-                cameraMovementFlags |= CameraMovement.RotateUp;
-
-            if (keysDown.Contains(Keys.PageDown) || keysDown.Contains(Keys.F))
-                cameraMovementFlags |= CameraMovement.RotateDown;
-
-            if (cameraMovementFlags != CameraMovement.None)
-                Mediator.Send(new Move3DViewRequest(cameraMovementFlags));
-
-            if (keysDown.Contains(Keys.OemOpenBrackets) || keysDown.Contains(Keys.OemCloseBrackets))
-            {
-                var zoomMagnitude = 0;
-
-                if (keysDown.Contains(Keys.OemOpenBrackets))
-                    zoomMagnitude += 100;
-                else if (keysDown.Contains(Keys.OemCloseBrackets))
-                    zoomMagnitude -= 100;
-
-                Mediator.Send(new Zoom3DViewRequest(zoomMagnitude));
-            }
-        }
-        
-        private void FreeCameraMovement(Keys[] keysDown)
-        {
-            var cameraMovementFlags = CameraMovement.None;
-
-            if (keysDown.Contains(Keys.A) || keysDown.Contains(Keys.Left))
-                cameraMovementFlags |= CameraMovement.PanLeft;
-
-            if (keysDown.Contains(Keys.D) || keysDown.Contains(Keys.Right))
-                cameraMovementFlags |= CameraMovement.PanRight;
-
-            if (keysDown.Contains(Keys.W) || keysDown.Contains(Keys.Up))
-                cameraMovementFlags |= CameraMovement.Forward;
-
-            if (keysDown.Contains(Keys.S) || keysDown.Contains(Keys.Down))
-                cameraMovementFlags |= CameraMovement.Backward;
-
-            if (keysDown.Contains(Keys.Q))
-                cameraMovementFlags |= CameraMovement.RotateLeft;
-
-            if (keysDown.Contains(Keys.E))
-                cameraMovementFlags |= CameraMovement.RotateRight;
-
-            if (keysDown.Contains(Keys.PageUp) || keysDown.Contains(Keys.R))
-                cameraMovementFlags |= CameraMovement.RotateUp;
-
-            if (keysDown.Contains(Keys.PageDown) || keysDown.Contains(Keys.F))
-                cameraMovementFlags |= CameraMovement.RotateDown;
-
-            if (cameraMovementFlags != CameraMovement.None)
-                Mediator.Send(new Move3DViewRequest(cameraMovementFlags));
-
-            if (keysDown.Contains(Keys.OemOpenBrackets) || keysDown.Contains(Keys.OemCloseBrackets))
-            {
-                var zoomMagnitude = 0;
-
-                if (keysDown.Contains(Keys.OemOpenBrackets))
-                    zoomMagnitude += 100;
-                else if (keysDown.Contains(Keys.OemCloseBrackets))
-                    zoomMagnitude -= 100;
-
-                Mediator.Send(new Zoom3DViewRequest(zoomMagnitude));
-            }
         }
     }
 }
