+++ DavidFidge.MonoGame.Core/Graphics/Camera/BaseCamera.cs
@@ -18,6 +18,10 @@ namespace DavidFidge.MonoGame.Core.Graphics.Camera
         protected int _fieldOfView;
         protected Vector3 _cameraPosition;
 
+        protected float _moveSpeed = 1f;
+        protected float _zoomSpeed = 0.1f;
+        protected float _rotateSpeed = 0.01f;
+
         public Matrix View { get; protected set; }
         public Matrix Projection { get; private set; }
